+++ p10-casee/p10-casee-api/src/main/java/com/pig4cloud/pig/casee/vo/ProjectStatisticsVO.java
@@ -11,10 +11,10 @@ public class ProjectStatisticsVO {
 	/**
 	 * 待接收统计
 	 */
-	private Integer pendingCount;
+	private Long pendingCount;
 
 	/**
 	 * 接收未处理统计
 	 */
-	private Integer notProcessedCount;
+	private Long notProcessedCount;
 }
